Add explaination to unmaintable controller
The functionality of both the category hub pages and the category-
restricted blogposts under the same URL leads to a hacky test
to see if the second parameter is numeric. The purpose of this test
is not obvious and the comment is needed to explain this so that the
code is more maintainable.